requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
mypy==0.971
mypy-extensions==0.4.3
Pillow==9.2.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
yarg==0.1.9
